+++ releases.md
@@ -2,6 +2,10 @@
     * Initialized the todo application
     * Added Dockerfile and docker-compose.yml file
     * Added redis
+    * Set up Go Gonic Gin route handler
 # v1.1.0
     * Added user repository
     * Added PostgreSQL database
+# v1.2.0
+    * Added User data handler using Go Gorilla Mux route handler
+    * Added Mux routes